TD1P1/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
Flask-mysqldb
Passlib
